@@ -1,9 +1,6 @@
 package com.example.dto;
 
-import javax.validation.constraints.Min;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Size;
+import jakarta.validation.constraints.*;
 
 import java.sql.Date;
 import java.time.LocalDateTime;
