@@ -1,7 +1,6 @@
 package com.example.dto;
 
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotBlank;
+import jakarta.validation.constraints.*;
 
 public class LoginRequest {
 
