@@ -1,9 +1,7 @@
 package com.example.dto;
 
-import javax.validation.Valid;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.Pattern;
-import javax.validation.constraints.Size;
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.*;
 
 import com.fasterxml.jackson.annotation.JsonAlias;
 
