@@ -4,6 +4,7 @@ import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -12,10 +13,10 @@ import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
-import com.example.dao.UserDAO;
 import com.example.dto.LoginRequest;
 import com.example.dto.UpdateCandidateDto;
 import com.example.dto.UserDto;
@@ -23,99 +24,139 @@ import com.example.entity.CandidateProfile;
 import com.example.entity.EmployerProfile;
 import com.example.entity.User;
 import com.example.enums.ExperienceLevel;
+import com.example.exception.UserEmailExistException;
+import com.example.exception.UserEmailNotFoundException;
+import com.example.repository.CandidateProfileRepository;
+import com.example.repository.EmployerProfileRepository;
+import com.example.repository.UserRepository;
 import com.example.security.JwtUtil;
 import com.example.service.UserService;
 
+/**
+ * Service implementation for User management using Spring Data JPA repositories.
+ * Handles user registration, authentication, profile management, and business logic.
+ */
 @Service
+@Transactional
 public class UserServiceImpl implements UserService {
 
     @Autowired
-    private BCryptPasswordEncoder passwordEncoder;
+    private PasswordEncoder passwordEncoder;
     
     @Autowired
-    private UserDAO userDAO;
+    private UserRepository userRepository;
+    
+    @Autowired
+    private CandidateProfileRepository candidateProfileRepository;
+    
+    @Autowired
+    private EmployerProfileRepository employerProfileRepository;
     
     @Autowired
     private AuthenticationManager authenticationManager;
     
-    private final JwtUtil jwtUtil = new JwtUtil();
+    @Autowired
+    private JwtUtil jwtUtil;
 
-    // Existing methods
+    // Basic CRUD operations
     @Override
+    @Transactional(readOnly = true)
     public List<User> findAllUsers() {
-        return userDAO.findAll();
+        return userRepository.findAll();
     }
 
     @Override
+    @Transactional(readOnly = true)
     public User findByUserId(int id) {
-        return userDAO.findById(id);
+        return userRepository.findById(id).orElse(null);
     }
 
     @Override
+    @Transactional(readOnly = true)
     public List<User> findByUserRole(String role) {
-        return userDAO.findByRole(role);
+        return userRepository.findByRole(role);
     }
 
     @Override
     public void saveUser(User user) {
-        userDAO.save(user);
+        if (userRepository.existsByEmail(user.getEmail())) {
+            throw new UserEmailExistException("Email already exists.");
+        }
+        
+        // Encode password
+        user.setPassword(passwordEncoder.encode(user.getPassword()));
+        
+        // Handle profile associations
+        if ("candidate".equalsIgnoreCase(user.getRole()) && user.getCandidateProfile() != null) {
+            user.getCandidateProfile().setUser(user);
+        }
+        if ("employer".equalsIgnoreCase(user.getRole()) && user.getEmployerProfile() != null) {
+            user.getEmployerProfile().setUser(user);
+        }
+        
+        userRepository.save(user);
     }
 
     @Override
     public void updateUser(User user) {
-        userDAO.update(user);
+        // Handle profile associations
+        if (user.getCandidateProfile() != null) {
+            user.getCandidateProfile().setUser(user);
+        }
+        if (user.getEmployerProfile() != null) {
+            user.getEmployerProfile().setUser(user);
+        }
+        
+        userRepository.save(user);
     }
 
     @Override
     public void deleteUser(int id) {
-        userDAO.delete(id);
+        userRepository.deleteById(id);
     }
 
     @Override
+    @Transactional(readOnly = true)
     public User findByemail(String email) {
-        return userDAO.findByEmail(email);
+        return userRepository.findByEmail(email)
+                .orElseThrow(() -> new UserEmailNotFoundException("No user found by the email: " + email));
     }
 
     @Override
     public void updatePassword(User user, String newPassword) {
         String encoded = passwordEncoder.encode(newPassword);
         user.setPassword(encoded);
-        userDAO.update(user); 
+        userRepository.save(user);
     }
 
-    // Business logic methods with ResponseEntity for role validation
+    // Authentication and business logic methods
     @Override
     public ResponseEntity<?> loginUser(LoginRequest loginRequest) {
         String email = loginRequest.getEmail();
         String password = loginRequest.getPassword();
         String role = loginRequest.getRole();
         
-        // 1. Authenticate the user by email and password (let exceptions bubble up)
+        // 1. Authenticate the user by email and password
         try {
             authenticationManager.authenticate(
                 new UsernamePasswordAuthenticationToken(email, password)
             );
         } catch (BadCredentialsException ex) {
-        	System.out.println("{\"message\": \"Invalid email or password!\"}");
             return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
                                  .body(Collections.singletonMap("message", "Invalid email or password"));
         }
         
         // 2. Fetch user from DB
         User dbUser = findByemail(email);
-        if (dbUser == null) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
-                                 .body(Collections.singletonMap("message", "User not found"));
-        }
-
-        // 3. Compare roles, assuming dbUser.getRole() returns "employer" or "candidate"
+        
+        // 3. Compare roles
         if (!dbUser.getRole().equalsIgnoreCase(role)) {
             return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
                                  .body(Collections.singletonMap("message", "Role does not match for this account"));
         }
 
         // 4. Generate JWT token
-        String token = jwtUtil.generateToken(dbUser.getEmail()); // or username if different
+        String token = jwtUtil.generateToken(dbUser.getEmail());
 
         Map<String, Object> response = new HashMap<>();
         response.put("status", "success");
@@ -129,21 +170,21 @@ public class UserServiceImpl implements UserService {
     public ResponseEntity<?> changeUserPassword(User request) {
         User dbUser = findByemail(request.getEmail());
         
-        if (dbUser == null) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND)
-                    .body(Collections.singletonMap("message", "Account not found with this email and role."));
-        }
-        
-        // Role validation - RETURN RESPONSE instead of throwing exception
+        // Role validation
         if (!dbUser.getRole().equalsIgnoreCase(request.getRole())) {
             return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
                     .body(Collections.singletonMap("message", "Role does not match for this account."));
         }
 
-        // Verify old password (let authentication exceptions bubble up)
-        authenticationManager.authenticate(
-            new UsernamePasswordAuthenticationToken(request.getEmail(), request.getOldPassword())
-        );
+        // Verify old password
+        try {
+            authenticationManager.authenticate(
+                new UsernamePasswordAuthenticationToken(request.getEmail(), request.getOldPassword())
+            );
+        } catch (BadCredentialsException ex) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                    .body(Collections.singletonMap("message", "Invalid old password."));
+        }
 
         // Update password
         updatePassword(dbUser, request.getNewPassword());
@@ -155,6 +196,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
+    @Transactional(readOnly = true)
     public ResponseEntity<?> getCurrentUserInfo(Authentication authentication) {
         User user = findByemail(authentication.getName());
 
@@ -174,7 +216,7 @@ public class UserServiceImpl implements UserService {
         user.setLastName(userDto.getLastName());
         user.setPhoneNumber(userDto.getPhoneNumber());
         user.setEmail(userDto.getEmail());
-        user.setPassword(userDto.getPassword()); // encode in DAO/save layer
+        user.setPassword(userDto.getPassword());
 
         if ("candidate".equalsIgnoreCase(userDto.getRole()) && userDto.getCandidateProfile() != null) {
             CandidateProfile cp = new CandidateProfile();
@@ -198,14 +240,10 @@ public class UserServiceImpl implements UserService {
         return ResponseEntity.ok(Collections.singletonMap("message", "User registered successfully"));
     }
 
-    
-    
     @Override
+    @Transactional(readOnly = true)
     public ResponseEntity<?> getUserProfile(Authentication authentication) {
         User user = findByemail(authentication.getName());
-        if (user == null) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("User not found");
-        }
 
         CandidateProfile candidateProfile = user.getCandidateProfile();
 
@@ -227,9 +265,6 @@ public class UserServiceImpl implements UserService {
     @Override
     public ResponseEntity<?> updateUserProfile(UpdateCandidateDto updateDto, Authentication authentication) {
         User user = findByemail(authentication.getName());
-        if (user == null) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("User not found");
-        }
 
         // Basic updates
         user.setFirstName(updateDto.getFirstName());
@@ -253,13 +288,12 @@ public class UserServiceImpl implements UserService {
         return ResponseEntity.ok(Collections.singletonMap("message", "Profile updated successfully"));
     }
 
-
-    // MOVED: Private utility methods from controllers
+    // Private utility methods
     private String mapExperienceToUi(ExperienceLevel experienceLevel) {
         if (experienceLevel == null) return null;
         
         switch (experienceLevel) {
-            case JUNIOR: return "1-3"; // representative value for JUNIOR range
+            case JUNIOR: return "1-3";
             case SENIOR: return "5-10";
             case LEAD: return "10+";
             default: return null;
@@ -272,7 +306,6 @@ public class UserServiceImpl implements UserService {
         uiValue = uiValue.trim();
         if ("10+".equals(uiValue)) return ExperienceLevel.LEAD;
         if ("5-10".equals(uiValue)) return ExperienceLevel.SENIOR;
-        // Map 0-1, 1-3, 3-5 all to JUNIOR
         return ExperienceLevel.JUNIOR;
     }
 }
